feat: move initHathorLib to helpers
